Add Windows support for QR code scanning (#207)
* UWP: Add support for QR code scanning

Signed-off-by: Alistair Francis <alistair@alistair23.me>

* Pages: ScanPage: Set the close button on Windows

A previous commit mistakenly removed the close button, even though the
scan page is not modal. This means there is no way to navigate away if
running on Windows 10. Revert this change to add a back button.

We need to use a modal page here as we need it to pop up over the entire
view.

Signed-off-by: Alistair Francis <alistair@alistair23.me>