Add RuPay
 Add RuPay (#1)

See https://en.wikipedia.org/wiki/RuPay
If this is not correct, please add RuPay as a card brand for the items cards. RuPay is a domestic credit and debit card. It is used by a lot of people in India.
For more info: https://www.rupay.co.in/who-we-are
https://en.wikipedia.org/wiki/RuPay