[PM-14843] Allow deletion of items in collections with manage permission
This change allows users to delete items if they have "manage" permission for any of the collections the item belongs to.

Previously, deletion was blocked if the item was in any collection where the user lacked "manage" permission. This is now corrected to allow deletion if the user has manage permissions for at least one collection containing the item.