[PM-15057] Update AndroidX Credentials to 1.5.0-alpha04
This commit updates the AndroidX Credentials dependency to version 1.5.0-alpha04, which includes new features and API changes to improve security for FIDO2 credential creation and retrieval. This introduces biometrics for both request types using the `BiometricPrompt` API. It also adds support for a new `DigitalCredential` type, as well as `RestoreCredential`, used to transfer the credential from the previous device to a new one.

Note: Versions `>= 1.5.0-alpha05` introduce breaking changes and bugs that are not yet suitable for production. This update targets `1.5.0-alpha04` while we wait for those issues to be resolved.