[PM-15176] Rename bundle and apk files to match applicationId and flavor
This change renames the bundle (AAB) and APK files generated during the build process. It uses the `applicationId` and `flavorName` to create more consistent and predictable filenames.

For APK files, the output filename is now set to `applicationId-flavorName.apk` for the "fdroid" flavor, `applicationId.apk` for the "standard" flavor, and the default output filename for other flavors.

For AAB files, a new task is registered for each variant to rename the bundle files. The task renames the AAB file to `applicationId-flavorName.aab` for the "fdroid" flavor and `applicationId.aab` for the "standard" flavor.

This change also removes the hardcoded renaming logic for specific variants and replaces it with a more generic approach that works for all variants.