[PM-12259] Defer passkey authentication until vault data is loaded
This commit refactors the FIDO2 Get Credentials request handling in `Fido2ProviderProcessorImpl` to use `takeUntilLoaded` and `fold` to collect `CipherView` data. This ensures querying for matching items is not performed until vault data for the active user is loaded.

We also defers the handling of FIDO2 assertion requests until the vault data is loaded. This ensures the selected item data is available before attempting authentication.

These changes prevent unexpected failures when a passkey authentication request is received while the application is not activeliy running.