Implement privileged app management for FIDO2
Introduces the ability to trust privileged applications for FIDO2 operations.

If a calling application is acting as a privileged application but is not yet trusted, the user will be given the option to trust the app and continue the operation, or to cancel the operation without trusting the application. If the application is trusted, subsequent requests from the application will be treated as a trusted application without user interaction.