[PM-19793] Migrate ZonedDateTimeSerializer to core module
Move ZonedDateTimeSerializer to the core module for reuse across modules.

The `providesJson` method has been removed from the network modules and moved to the `CoreModule`.

All usages of `providesJson` from the old locations were updated to use the `CoreModule`.

The `ZonedDateTimeSerializerTest` class was also moved from `app` to `core` module.