Bump authenticatorbridge to 1.0.1
This commit updates the `authenticatorbridge` library to version 1.0.1.

This version includes a bug fix that sets the `AccountSyncState` to `SyncNotEnabled` after binding to the service if the previous state was `AppNotInstalled`.

The `authenticatorbridge-1.0.0-release.aar` file has been replaced with `authenticatorbridge-1.0.1-release.aar` in both the `app` and `authenticator` modules.