[PM-21782] Pass `encryptedFor` to cipher functions
This commit updates several functions to accept an `encryptedFor` parameter, which specifies the user ID for whom the cipher is encrypted. This property is used by the server to verify the cipher is encrypted by the correct user. If verification fails the server responds with an appropriate error message.

This change affects the following:
- `toEncryptedNetworkCipher` and `toEncryptedNetworkCipherResponse` extension functions for `Cipher` now require an `encryptedFor` parameter.
- `CipherJsonRequest` and `SyncResponseJson.Cipher` now include an `encryptedFor` field.
- SDK functions like `encryptCipher` and `decryptFile` have been updated to align with these changes.

Additionally, this update includes the following SDK related changes:
- Adding the `encryptedFor` related logic (https://github.com/bitwarden/sdk-internal/pull/278)
- Update in protocol and objects naming to have ``{foo}Client[Protocol]`` instead of `Client{foo}[Protocol]` (https://github.com/bitwarden/sdk-internal/pull/224)
- Update attachments decryption to use `AttachmentView` instead of `Attachment` (https://github.com/bitwarden/sdk-internal/pull/255)