[PM-19108] Update passkey prompt for unrecognized browser
The dialog title for the trust privileged add prompt is changed from "An error has occurred" to "Unrecognized browser".

The message for the passkey operation failure due to an untrusted browser is updated to state that the browser is "not recognized" instead of "not trusted", and specifies that trusting will add it to the list of "locally trusted applications".

A new string resource "unrecognized_browser" is added.