[PM-22998] Migrate `isBuildVersionBelow` to core module
Replaces all instances of `isBuildVersionBelow` with the more appropriately named `isBuildVersionAtLeast` for checking Android SDK versions. This change improves code readability and aligns with the common practice of checking if a version is at least a certain level.

The new `isBuildVersionAtLeast` function is also annotated with `@ChecksSdkIntAtLeast` for better static analysis.

The `authenticatorbridge` module now includes `core` as an implementation dependency.