[BWA-162] Add `getPackageInstallationSourceOrNull` to `BitwardenPackageManager`
This commit introduces a new method,
`getPackageInstallationSourceOrNull`, to the `BitwardenPackageManager`
interface and its implementation, `BitwardenPackageManagerImpl`.

This method allows retrieval of the installation source of a given
package. It returns the package name of the installer application or
null if the information is unavailable (e.g., system app, app not
installed, or installer uninstalled).

The implementation leverages `PackageManager.getInstallSourceInfo`
available from Android R (API level 30) and is annotated with
`@RequiresApi(Build.VERSION_CODES.R)`.

Unit tests have been added to verify the functionality of
`getPackageInstallationSourceOrNull`.