[PM-23605] Add decryptCipherListWithFailures to VaultSdkSource
This commit introduces a new method `decryptCipherListWithFailures` to the `VaultSdkSource` interface and its implementation `VaultSdkSourceImpl`. This method decrypts a list of ciphers and returns a `DecryptCipherListResult` which includes both successfully decrypted ciphers and references to ciphers that failed to decrypt.

The existing methods `decryptCipherListCollection` and `decryptCipherList` will be deprecated in favor of the new method.

Unit tests have been added to verify the functionality of `decryptCipherListWithFailures`, including success and failure scenarios.