[PM-23608] Add generateTotp overload with CipherListView parameter
The `VaultSdkSource` and `VaultSdkSourceImpl` classes are updated to include a new overload for the `generateTotp` method. This overload accepts a `CipherListView` parameter instead of a `String` totp parameter.

This change also updates the corresponding unit tests to specify argument names.

Additionally, a new utility file `CipherListViewUtil.kt` is introduced to create mock `CipherListView` objects for testing purposes.