[PM-23665] Refactor FIDO2 credential discovery
This commit refactors the FIDO2 credential discovery process in `BitwardenCredentialManagerImpl`. Instead of relying on decrypted `CipherListViews`, it now uses `vaultSdkSource.silentlyDiscoverCredentials` to find matching credentials based on the `relyingPartyId` from the assertion options.

The discovered credentials are then filtered by `allowedCredentials` if provided in the assertion options.

Additionally, some related test cases in `BitwardenCredentialManagerTest` were updated to reflect these changes and ensure proper handling of new scenarios, such as when `relyingPartyId` is null.