[PM-26095] Add account selection screen for Credential Exchange
This commit introduces a new screen that allows users to select an account when importing items via the Credential Exchange Framework (CXF).

When a Credential Exchange import request is received the user is prompted to confirm which account to export items from.

Should the user choose to close the screen, the export process will be cancelled and the user returned to the requesting application.

Key changes include:
- Added `SelectAccountScreen.kt` for the UI of the account selection.
- Added `SelectAccountViewModel.kt` to manage the state and logic for account selection.
- Added `ExportItemsScaffold` a reusable scaffold for the export items screens.
- Included `AccountSelectionListItem.kt` to model the items displayed in the account list.
- Implemented `SelectAccountHandlers.kt` to manage user interactions on the screen.
- Added unit tests for the new screen (`SelectAccountScreenTest.kt`) and view model (`SelectAccountViewModelTest.kt`).
- Added new string resources for the account selection screen.
- Updated `BitwardenComposeTest.kt` and `LocalManagerProvider.kt` to include `CredentialExchangeCompletionManager`.