[PM-26094] Update Credential Manager library and remove stubs
This commit updates the AndroidX Credential Manager library to `1.6.0-beta01` and the Credential Manager Provider Events library to `1.0.0-alpha03`.

Additionally, this removes stub implementations for `androidx.credentials.providerevents` that are no longer needed with the updated library versions.

The `CredentialTypes` constants used in `ImportItemsViewModel` and its corresponding test have also been updated to reflect changes in the newer library version.