[PM-26804] Clear password input after verification in `VerifyPasswordViewModel`
This commit addresses a security concern where the password input in the `VerifyPasswordViewModel` was not being cleared after successful verification.

To resolve this, the password input field is now cleared from the view model's state upon successful password validation or vault unlock. The `input` field in `VerifyPasswordState` is also annotated with `@IgnoredOnParcel` to prevent it from being saved to the instance state, further enhancing security.

Additionally, the `onCleared()` method is overridden to explicitly clear the password input, as a temporary workaround for a suspected OS-level memory leak. Corresponding unit tests have been updated to verify that the input is cleared after successful verification.